+++ js/lib/flynn/flynn3D.js
@@ -8,12 +8,14 @@ Flynn._3DCamera = Class.extend({
 });
 
 Flynn._3DMesh = Class.extend({
-    init: function(name, vertices, lines, color){
+    init: function(name, vertices, lines, color, custom_palette){
         this.name = name;
-        this.color = color;
         this.Vertices = vertices;
-        this.ProjectedVertices = new Array(vertices.length);
         this.Lines = lines;
+        this.color = color;
+        this.custom_palette = typeof(custom_palette)==='undefined' ? null : custom_palette;
+
+        this.ProjectedVertices = new Array(vertices.length);
         this.Rotation = BABYLON.Vector3.Zero();
         this.Position = BABYLON.Vector3.Zero();
     },
@@ -49,10 +51,66 @@ Flynn._3DMeshCube = Flynn._3DMesh.extend({
     },
 });
 
+Flynn._3DMeshFromPoints = Flynn._3DMesh.extend({
+    init: function(name, points, scale, color){
+        var i;
+        var vertices = [];
+        var lines = [];
+        var num_vertices = 0;
+        for(i=0; i<points.length; i+=2){
+            if(points[i] == Flynn.PEN_COMMAND){
+                if(points[i+1] == Flynn.PEN_UP){
+                    lines.push(Flynn.PEN_UP);
+                }
+                continue;
+            }
+            vertices.push( new BABYLON.Vector3(
+                points[i]*scale,
+                0,
+                points[i+1]*scale));
+            lines.push(num_vertices++);
+        }
+        this._super(name, vertices, lines, color);
+    }
+});
+
+Flynn._3DMeshPlate = Flynn._3DMesh.extend({
+    init: function(name, width, depth, color){
+        var loc_x = width/2;
+        var loc_z = depth/2;
+        var vertices = [
+            new BABYLON.Vector3( loc_x, 0,  loc_z),
+            new BABYLON.Vector3(-loc_x, 0,  loc_z),
+            new BABYLON.Vector3(-loc_x, 0, -loc_z),
+            new BABYLON.Vector3( loc_x, 0, -loc_z),
+        ];
+
+        var lines =[0, 1, 2, 3, 0];
+
+        this._super(name, vertices, lines, color);
+    },
+});
+
 Flynn._3DRenderer = Class.extend({
     VERTEX_SIZE: 2,
 
     init: function(opts){
+        // Options:
+        //    width: (int) The width of the display canvas
+        //    height: (int) The height of the display canvas
+        //    fog_distance: Either null for no fog, or an object of the
+        //        form {near:float, far:float} 
+        //           near: The distance from camera beyond which meshes
+        //                 will begin to fade out.
+        //           far:  The distance from camera beyond which meshes
+        //                 will fade out entirely (not be drawn)
+        //    enable_vertices: (boolean) Set true to draw vertices as points.
+        //    enable_lines: (boolean) Set true to draw lines (vectors) between
+        //        vertices (per the mesh.Lines array)
+        //    enable_distance_order: (boolean) Set true to sort meshes by their
+        //        distance from the camera and draw in order from farthest 
+        //        to closest.
+        //         
         opts                       = opts                    || {};
         this.width                 = opts.width              || Flynn.mcp.canvasWidth;
         this.height                = opts.height             || Flynn.mcp.canvasHeight;
@@ -82,6 +140,7 @@ Flynn._3DRenderer = Class.extend({
                 0.78, this.width / this.height, 0.01, 1.0);
 
         var color;
+        var dim_factor = 0;
         var visible;
         var draw_list = [];
         for (var index = 0; index < meshes.length; index++) {
@@ -106,7 +165,7 @@ Flynn._3DRenderer = Class.extend({
                     visible = false;
                 }
                 else if (distance > this.fog_distance.near){
-                    var dim_factor = -((distance - this.fog_distance.near) / 
+                    dim_factor = -((distance - this.fog_distance.near) / 
                         (this.fog_distance.far - this.fog_distance.near));
                     color = Flynn.Util.shadeColor(cMesh.color, dim_factor);
                 }
@@ -139,18 +198,24 @@ Flynn._3DRenderer = Class.extend({
                             this.VERTEX_SIZE);
                     }
                 }
-                draw_list.push({mesh_index:index, distance:distance, color:color});
+                draw_list.push({mesh_index:index, distance:distance, color:color, dim_factor:dim_factor});
             }
         }
 
-        // Sort draw order by distance
-        if(this.enable_distance_order){
-            draw_list.sort(function(a,b){return b.distance - a.distance;});
-        }
 
         if(this.enable_lines){
+            
+            // Sort draw order by distance
+            if(this.enable_distance_order){
+                draw_list.sort(function(a,b){return b.distance - a.distance;});
+            }
+
+            //------------------------
+            // Draw vectors
+            //------------------------
             for(var i=0; i<draw_list.length; i++){
                 var target = draw_list[i];
+                color = target.color;
                 cMesh = meshes[target.mesh_index];
                 var lines = cMesh.Lines;
                 var vertices = cMesh.ProjectedVertices;
@@ -159,18 +224,39 @@ Flynn._3DRenderer = Class.extend({
                 for(var index_lines = 0; index_lines < lines.length; index_lines++){
                     var index_vertex = lines[index_lines];
                     if (index_vertex == Flynn.PEN_UP){
+                        // Pen up
                         ctx.vectorEnd();
+                        pen_up = true;
+                        continue;
+                    }
+                    if (index_vertex >= Flynn.PEN_COLOR0){
+                        // Switch colors
+                        var color_index = index_vertex-Flynn.PEN_COLOR0;
+                        if(cMesh.custom_palette == null){
+                            color = Flynn.ColorsOrdered[color_index];
+                        }
+                        else{
+                            color = cMesh.custom_palette[color_index];
+                        }
+                        if(target.dim_factor != 0){
+                            color = Flynn.Util.shadeColor(color, target.dim_factor);
+                        }
+                        if(started){
+                            ctx.vectorEnd();
+                        }
                         pen_up = true;
                         continue;
                     }
                     var vertex = vertices[index_vertex];
                     if(pen_up){
-                        ctx.vectorStart(target.color, false, false);
+                        // Start line
+                        ctx.vectorStart(color, false, false);
                         ctx.vectorMoveTo(vertex.x, vertex.y);
                         pen_up = false;
                         started = true;
                     }
                     else{
+                        // Draw line
                         ctx.vectorLineTo(vertex.x, vertex.y);
                     }
                 }