+++ CHANGELOG.rst
@@ -33,6 +33,7 @@ Changed
 - Engine
 
   - Improve shape of ")" and "(" in main vector font.
+  - Add characters to main vector font: "{}|~"
 
 3.7.3 - 2018-4-10
 -----------------